+++ blocklets/inscription/api/routes/auth/index.js
@@ -1,10 +1,12 @@
 const { handlers } = require('../../libs/auth');
 const deployContractHandler = require('./deploy-contract');
 const recordMessageHandler = require('./record-message');
+const verifyBlockletOwnerHandler = require('./verify-blocklet-owner');
 
 module.exports = {
   init(app) {
     handlers.attach({ app, ...deployContractHandler });
     handlers.attach({ app, ...recordMessageHandler });
+    handlers.attach({ app, ...verifyBlockletOwnerHandler });
   },
 };